my_astro_bot/requirements.txt
aiogram==3.17.0
ephem==4.1.6
geopy>=2.4.0
timezonefinder>=6.2.0
pydantic==2.10.6
python-dotenv==1.0.1
pytz==2025.1
python-dateutil==2.9.0.post0
requests==2.32.3
matplotlib>=3.10
numpy>=1.23
